% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WallomicsData-package.R
\docType{data}
\name{Altitude_Cluster}
\alias{Altitude_Cluster}
\title{Altitude Cluster}
\format{
A factor with 3 levels.
}
\source{
\doi{10.3390/cells9102249}
}
\usage{
data("Altitude_Cluster")
}
\description{
The \strong{Altitude Cluster factor} identifies the environment height
from which is originated a given plant from the sample under study, either high altitude (denoted \emph{High}),
moderate altitude (\emph{Low}) or the reference group's environment height (\emph{Col}, the lowest of all).
}
\examples{
# Load the data
data("Altitude_Cluster")

# Count how many samples are in each group
table(Altitude_Cluster)

}
\keyword{datasets}
