% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WallomicsData-package.R
\docType{data}
\name{Genetic_Cluster}
\alias{Genetic_Cluster}
\title{Genetic Cluster}
\format{
A factor with 3 levels.
}
\source{
\doi{10.3390/cells9102249}
}
\usage{
data("Genetic_Cluster")
}
\description{
The \strong{Genetic Cluster factor} identifies the genetic group
from which comes from the studied sample, either \strong{Genetics Cluster 1} (constitued of \emph{Grip} and \emph{Roch} genotypes),
\strong{Genetics Cluster 2} (\emph{Hern} and \emph{Hosp} genotypes) or \strong{Genetics Cluster 3} (\emph{Col} genotype).
See \link{Ecotype} for more information on genotypes.
}
\examples{
# Load the data
data("Genetic_Cluster")

# Count how many samples are in each group
table(Genetic_Cluster)

}
\keyword{datasets}
