% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WallomicsData-package.R
\docType{data}
\name{Temperature}
\alias{Temperature}
\title{Temperature}
\format{
A factor with 2 levels.
}
\source{
\doi{10.3390/cells9102249}
}
\usage{
data("Temperature")
}
\description{
The \strong{Temperature factor} identifies the temperature at which the studied sample was exposed
all along its growth, either 22°C (optimal condition) or 15°C (high altitude condition).
}
\examples{
# Load the data
data("Temperature")

# Count how many samples are in each group
table(Temperature)

}
\keyword{datasets}
