% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WallomicsData-package.R
\docType{package}
\name{WallomicsData-package}
\alias{WallomicsData}
\alias{WallomicsData-package}
\title{WallomicsData: Datasets for Multi-Omics Integration in a Plant Abiotic Stress Context}
\description{
Datasets from the WallOmics project. Contains phenomics, metabolomics, proteomics and transcriptomics data collected from two organs of five ecotypes of the model plant Arabidopsis thaliana exposed to two temperature growth conditions. Exploratory and integrative analyses of these data are presented in Durufle et al (2020) <doi:10.1093/bib/bbaa166> and Durufle et al (2020) <doi:10.3390/cells9102249>.
}
\author{
\strong{Maintainer}: Sébastien Déjean \email{sebastien.dejean@math.univ-toulouse.fr}

Authors:
\itemize{
  \item Harold Duruflé \email{harold.durufle@inrae.fr}
  \item Aurélie Mercadié \email{aurelie.mercadie@inrae.fr}
}

}
\keyword{internal}
